+++ Lib/SceneScrollerClass.js
@@ -1,0 +1,582 @@
+import { ModuleName, ssc } from "../ss-initialize.js";
+import * as Viewport from "./ViewportClass.js";
+import { ScrollerToken } from "./TokenClass.js";
+
+/**
+ * A class that will be publicly available containing a schema for
+ * viewport, tile and token flags.
+ */
+export class SCSC_Flag_Schema {
+    /* Default keys and flag values */
+    static viewportFlags = {
+        SubSceneTokenData: [],  // Stored as JSON strings
+        ActiveTokenID: "",
+        ActiveScene: ""
+    }
+
+    static compendiumSceneFlags = {
+        SubSceneChildren: [],
+        Bounds: {},
+        Coords: {},
+        Polygon : []
+    }
+
+    static subSceneChildrenFlags = {
+        ChildrenSceneUUIDs: [],
+        ChildCoords: {},
+        Tile: {},
+        Bounds: {}
+    }
+
+    static tokenFlags = {
+        CurrentSubScene: "",
+        InSubSceneLoc: {},
+    }
+}
+
+/**
+ * A class that caches placeables and handles CRUD operations.
+ */
+export class SceneScroller_Cache {
+    constructor() {
+        this.viewportFlags = Object.keys(SCSC_Flag_Schema.viewportFlags);
+        this.subSceneChildrenFlags = Object.keys(SCSC_Flag_Schema.subSceneChildrenFlags);
+        this.tokenFlags = Object.keys(SCSC_Flag_Schema.tokenFlags);
+        this.compendiumFlags = Object.keys(SCSC_Flag_Schema.compendiumSceneFlags);
+        this.viewport = {};
+        this.subScenes = new Map();
+        this.tokens = new Map();
+        this.walls = new Map();
+        this.drawings = new Map();
+        this.lights = new Map();
+        this.notes = new Map();
+        this.sounds = new Map();
+        this.templates = new Map();
+        this.tiles = new Map();
+        this.tokens = new Map();
+        this.compendiumSources = new Map();
+        this.sprites = new Map();
+        this.selTokenApp = null;
+
+        this._initialize();
+    }
+
+    /*************************************************************************************/
+    /* viewport CRUD operations */
+    /*************************************************************************************/
+
+    /**
+     * A getter to fetch the active sub-scene to display.  Displayed sub-scene can differ by user.
+     */
+    get activeSceneUUID() {
+        if ( this.viewport[this.viewportFlags[2]] !== undefined ) return this.viewport[this.viewportFlags[2]];
+    }
+
+    /**
+     * @param {string} uuid The UUID of the sub-scene to be cached.
+     */
+    cacheactiveSceneUUID(uuid) {
+        this.viewport[this.viewportFlags[2]] = uuid;
+    }
+
+    /**
+     * Sets the supplied token ID as active in the cache and in the canvas flags.
+     * @param {string} tokenID The id of the token in the cache
+     */
+    async setActiveToken(tokenID) {
+        this.viewport[this.viewportFlags[1]] = tokenID;
+        await canvas.scene.setFlag(ModuleName, this.viewportFlags[1], tokenID);
+    }
+
+    /**
+     * A getter to fetch the active token ID from the cache.
+     */
+    get activeTokenID() {
+        return this.viewport[this.viewportFlags[1]];
+    }
+
+    /*************************************************************************************/
+    /* subScenes CRUD operations */
+    /*************************************************************************************/
+
+    /** A getter to fetch an array of all the active sub-scenes from the cache. */
+    get allSubScenes() {
+        return [...new Set(this.subScenes.values())];
+    }
+
+    /**
+     * Get a tile instance from the cache
+     * @param {string} id The tile.ID
+     * @returns {object}    A foundry Tile instance.
+     */
+    getSubSceneTile(id) {
+        return this.subScenes.get(id);
+    }
+
+    /**
+     * Save a tile instance to the cache
+     * @param {string} id the tile.ID
+     * @param {object} tile A foundry Tile instance
+     */
+    setSubSceneCache(id, tile) {
+        this.subScenes.set(id, tile);
+    }
+
+    /** 
+     * Check if the cache contains a tile instance
+     * @param {string}  id  The tile.ID or the UUID
+     * @returns {boolean}   True if it does, false otherwise
+     */
+    hasSubSceneInCache(id) {
+        return this.subScenes.has(id);
+    }
+
+    /**
+     * A getter to fetch a bounds object (data) of the current active sub-scene stored in the cache
+     */
+    get ActiveBounds() {
+        return this.subScenes.get(this.activeSceneUUID)[this.subSceneChildrenFlags[3]];
+    }
+
+    /**
+     * A getter to fetch an array of sub-scene UUID's for all the child sub-scenes of the
+     * current active sub-scene stored in the cache
+     */
+    get ActiveChildrenUUIDs() {
+        const activeSubSceneUuid = this.subScenes.get(this.activeSceneUUID).compendiumSubSceneUUID;
+        const activeSubSceneSource = this.compendiumSources.get(activeSubSceneUuid);
+        const subSceneFlagsArr =  activeSubSceneSource.getFlag("scene-scroller-maker", this.compendiumFlags[0]);
+        return subSceneFlagsArr.map(s => {
+            return s[this.subSceneChildrenFlags[0]];
+        })
+    }
+
+    /**
+     * A function that returns an array of children uuids for a given parent uuid (the parent sub-scene).
+     * @param {string} uuid The UUID of the sub-scene from which we want the children uuids.
+     * @returns {array<string>}
+     */
+    childrenUuids(uuid) {
+        const activeSubSceneSource = this.compendiumSources.get(uuid);
+        const subSceneFlagsArr =  activeSubSceneSource.getFlag("scene-scroller-maker", this.compendiumFlags[0]);
+        return subSceneFlagsArr.map(s => {
+            return s[this.subSceneChildrenFlags[0]];
+        })
+    }
+
+    get ActiveChildrenFlags() {
+        const activeSubSceneUuid = this.subScenes.get(this.activeSceneUUID).compendiumSubSceneUUID;
+        const activeSubSceneSource = this.compendiumSources.get(activeSubSceneUuid);
+        return activeSubSceneSource.getFlag("scene-scroller-maker", this.compendiumFlags[0]) 
+    }
+    
+    /*************************************************************************************/
+    /* tokens CRUD operations */
+    /*************************************************************************************/
+
+    /**
+     * Add a Token to the cache.
+     * If user is GM, then token is added to an array of Tokens present in the scene, as flags to the scene
+     * If user is GM, then flag added to scene setting this Token as the active token.
+     * @param {object} token A Foundry Token instance
+     */
+    async cacheToken(token) {
+        this.tokens.set(token.document.ss_id, token);
+        if ( game.user.isGM ) {
+            const data = token.document.toObject();
+            data.ss_id = token.document.ss_id;
+            const currArr = canvas.scene.getFlag(ModuleName, this.viewportFlags[0]) || [];
+            currArr.push(JSON.stringify(data));
+            await canvas.scene.setFlag(ModuleName, this.viewportFlags[0], currArr);
+            await canvas.scene.setFlag(ModuleName, this.viewportFlags[1], token.document.ss_id);
+        }
+    }
+
+    /**
+     * Gets a token from cache.
+     * @param {string} id A Token instance ID
+     * @returns {object}    A Foundry Token instance
+     */
+    getToken(id) {
+        return this.tokens.get(id);
+    }
+
+    /**
+     * A getter to obtain an array of all Token instances in the cache
+     */
+    get getAllTokens() {
+        return [...new Set(this.tokens.values())];
+    }
+
+    /**
+     * Deletes a Token from cache
+     * If user is GM, deletes Token from array stored in scene flags
+     * If user is GM, sets a new active token, if necessary.
+     * @param {object} token A Foundry Token instance
+     */
+    async deleteToken(token) {
+        this.tokens.delete(token.id);
+        if ( game.user.isGM ) {
+            const currArr = canvas.scene.getFlag(ModuleName, this.viewportFlags[0]) || [];
+            const newArr = currArr.filter(t => !t.includes(JSON.stringify(token.id)));
+            await canvas.scene.setFlag(ModuleName, this.viewportFlags[0], newArr);
+            // If this happened to also be the active token, then pick the first from newArr to make it the active token
+            const currTokID = canvas.scene.getFlag(ModuleName, this.viewportFlags[1]);
+            if ( currTokID === token.id ) {
+                const newCurr = newArr[0]?.id || "";
+                await canvas.scene.setFlag(ModuleName, this.viewportFlags[1], newCurr);
+            }
+        }
+    }
+
+    /**
+     * Updates the flags stored in Token instances.  To make token movements persist.
+     * @param {object} token A Foundry Token instance
+     * @param {object} loc Coordinates of the token relative to a sub-scene top left corner.  {x: <number> y: <number>}
+     * @param {string} uuid *optional* The compendium scene UUID of the sub-scene the token occupies.
+     */
+    async updateTokenFlags(token, loc, uuid = null) {
+
+        // Tokens are local memory only.  Not in db.  Can't use setFlag
+        token.document.updateSource({
+            [`flags.${ModuleName}.${ssc.tokenFlags[1]}`] : {x: loc.x, y: loc.y}
+        });
+
+        if ( uuid !== null ) {
+            token.document.updateSource({
+                [`flags.${ModuleName}.${ssc.tokenFlags[0]}`] : uuid
+            });
+        }
+
+        await this.deleteToken(token);
+        await this.cacheToken(token);
+    }
+
+    /**
+     * Gets a UUID defining what sub-scene the token occupies.
+     * @param {object} token A Foundry Token instance
+     * @returns {string} The UUID of the compendium sub-scene occupied by the Token
+     */
+    tokenCurrentSubScene(token) {
+        return token.document.getFlag(ModuleName, this.tokenFlags[0]);
+    }
+
+    /**
+     * Gets an object defining the coordinates of the token relative to the sub-scene it occupies.
+     * @param {object} token A Foundry Token instance
+     * @returns {object} {x: <number>, y: <number>}  Coordinates relative to sub-scene top left corner.
+     */
+    tokenCurrentLoc(token) {
+        return token.document.getFlag(ModuleName, this.tokenFlags[1]);
+    }
+
+    /*************************************************************************************/
+    /* Placeables CRUD Methods */
+    /*************************************************************************************/
+
+    /**
+     * Adds a wall placeable to the cache
+     * @param {object} wall A Foundry Wall instance
+     */
+    addWall(wall) {
+        if ( this.walls.has(wall.id) ) {
+            // Duplicate wall.  Update the wall to add parent ID existing wall.
+            const existingWall = this.walls.get(wall.id);
+            const otherParentID = existingWall.parentSubScene[0];
+            wall.parentSubScene.push(otherParentID);
+            this.walls.set(wall.id, wall);
+        } else this.walls.set(wall.id, wall);
+    }
+
+    /**
+     * 
+     * @param {object} wall A Foundry Wall instance
+     * @param {string} id   The ID of the Foundry Tile instance that is triggering the removal of the wall.
+     */
+    removeWall(wall, id) {
+        if ( wall.parentSubScene.length > 1 ) {
+            const filtered = wall.parentSubScene.filter(w => w.parentSubScene !== id);
+            wall.parentSubScene = filtered;
+        } else this.walls.delete(wall.id);
+    }
+
+    /**
+     * Adds a light placeable to the cache
+     * @param {object} light A Foundry Light instance
+     */
+    addLight(light) {
+        this.lights.set(light.id, light);
+    }
+
+    /**
+     * 
+     * @param {object} light A Foundry Light instance
+     */
+    removeLight(light) {
+        this.lights.delete(light.id);
+    }
+
+    /**
+     * Adds a note placeable to the cache
+     * @param {object} note A Foundry Note instance
+     */
+    addNote(note) {
+        this.notes.set(note.id, note);
+    }
+
+    /**
+     * 
+     * @param {ojbect} note A Foundry Note instance
+     */
+    removeNote(note) {
+        this.notes.delete(note.id);
+    }
+
+    /**
+     * Adds a sound placeable to the cache
+     * @param {object} sound A Foundry Sound instance
+     */
+    addSound(sound) {
+        this.sounds.set(sound.id, sound);
+    }
+
+    /**
+     * 
+     * @param {object} sound A Foundry Sound instance
+     */
+    removeSound(sound) {
+        this.sounds.delete(sound.id);
+    }
+
+    /**
+     * Adds a template placeable to the cache
+     * @param {object} template A Foundry Template instance
+     */
+    addTemplate(template) {
+        this.templates.set(template.id, template);
+    }
+
+    /**
+     * 
+     * @param {object} template A Foundry Template instance
+     */
+    removeTemplate(template) {
+        this.templates.delete(template.id);
+    }
+
+    /**
+     * Adds a tile placeable to the cache
+     * @param {object} tile A Foundry Tile instance
+     */
+    addTile(tile) {
+        this.tiles.set(tile.id, tile);
+    }
+
+    /**
+     * 
+     * @param {object} tile A Foundry Tile instance
+     */
+    removeTile(tile) {
+        this.tiles.delete(tile.id);
+    }
+
+    /**
+     * Adds a drawing placeable to the cache
+     * @param {object} drawing A Foundry Drawing instance
+     */
+    addDrawing(drawing) {
+        this.drawings.set(drawing.id, drawing);
+    }
+
+    /**
+     * 
+     * @param {object} drawing A Foundry Drawing instance
+     */
+    removeDrawing(drawing) {
+        this.drawings.delete(drawing.id);
+    }
+
+    /*************************************************************************************/
+    /* Compendium Source CRUD Methods */
+    /*************************************************************************************/
+
+    cacheCompendiumSource(id, source) {
+        this.compendiumSources.set(id, source);
+    }
+
+    compendiumSourceFromCache(uuid) {
+        return this.compendiumSources.get(uuid);
+    }
+
+    /*************************************************************************************/
+    /* sprites CRUD Methods */
+    /*************************************************************************************/
+
+    cacheSubSceneSprite(id, sprite) {
+        this.sprites.set(id, sprite);
+    }
+
+    spriteFromCache(uuid) {
+        return this.sprites.get(uuid);
+    }
+
+    /*************************************************************************************/
+    /* NON-CRUD Methods */
+    /*************************************************************************************/
+
+    /**
+     * Populates a SceneScroller_Cache instance with data on creation.
+     */
+    async _initialize() {
+        for (const k of this.viewportFlags) {
+            this.viewport[k] = canvas.scene.getFlag(ModuleName, k)
+        }
+
+        // Create full tokens from stored tokenData.
+        if ( !Viewport.isScrollerScene() ) return;
+        const tokArray = canvas.scene.getFlag(ModuleName, this.viewportFlags[0]);
+        for (const tok of tokArray) {
+            const data = JSON.parse(tok);
+            // Creating a new TokenDocument will mutate data and scrub out .ss_id
+            const ssId = data.ss_id;
+            const doc = new CONFIG.Token.documentClass(data, {parent: canvas.scene});
+            const token = new ScrollerToken(doc);
+            doc.ss_id = ssId;
+            doc._object = token
+            this.tokens.set(token.id, token);
+        }
+
+        // Cache the active token's active sub-scene
+        const activeTokenID = canvas.scene.getFlag(ModuleName, this.viewportFlags[1]);
+        const activeToken = this.tokens.get(activeTokenID);
+        const activeTokenScene = activeToken.document.getFlag(ModuleName, this.tokenFlags[0]);
+        this.cacheactiveSceneUUID(activeTokenScene);
+    }
+}
+
+/**
+     * Queries a scene in a compendium to copy all placeables into the cache.
+     * All copied placeables reference the parent Tile/Scene
+     * @param {object} scene An compendium scene
+     * @param {object} tile A Foundry Tile instance
+     */
+export function cacheInScenePlaceables(scene, tile) {
+    const placeables = ["drawings", "lights", "notes", "sounds", "templates", "tiles", "tokens", "walls"];
+    const pDict = {
+        drawings: {
+            doc : (data) => new DrawingDocument(data, {parent: canvas.scene}),
+            p: (doc) => new Drawing(doc),
+            cache: (d) => ssc.addDrawing(d)
+        },
+        lights : {
+            doc : (data) => new AmbientLightDocument(data, {parent: canvas.scene}),
+            p: (doc) => new AmbientLight(doc),
+            cache: (l) => ssc.addLight(l)
+        },
+        notes : {
+            doc : (data) => new NoteDocument(data, {parent: canvas.scene}),
+            p: (doc) => new Note(doc),
+            cache: (n) => ssc.addNote(n)
+        },
+        sounds: {
+            doc : (data) => new AmbientSoundDocument(data, {parent: canvas.scene}),
+            p: (doc) => new AmbientSound(doc),
+            cache: (s) => ssc.addSound(s)
+        },
+        templates: {
+            doc : (data) => new MeasuredTemplateDocument(data, {parent: canvas.scene}),
+            p: (doc) => new MeasuredTemplate(doc),
+            cache: (t) => ssc.addTemplate(t)
+        },
+        tiles: {
+            doc: (data) => new TileDocument(data, {parent: canvas.scene}),
+            p: (doc) => {return doc.object},
+            cache: (t) => ssc.addTile(t)
+        },
+        tokens: {
+            doc: (data) => new TokenDocument(data, {parent: canvas.scene}),
+            p: (doc) => new Token(doc),
+            cache: (t) => ssc.cacheToken(t)
+        },
+        walls: {
+            doc: (data) => new WallDocument(data, {parent: canvas.scene}),
+            p: (doc) => new Wall(doc),
+            cache: (w) => ssc.addWall(w)
+        }
+    }
+
+    for (const placeable of placeables) {
+        scene[placeable].forEach(p=> {
+            const data = p.toObject();
+            const doc = pDict[placeable].doc(data);
+            const e = pDict[placeable].p(doc);
+            e.parentSubScene = [tile.id]
+            pDict[placeable].cache(e)
+        })
+    }
+}
+
+/** Given an array of compendium scene UUID's, create Froundry Tiles for each, cache them and recreate the placeables.
+ * @param {string|array.<string>}   uuids   The array of uuid strings
+ */
+export async function cacheSubScene(uuids) {
+    if ( !Array.isArray(uuids) ) uuids = [uuids];
+
+    for (const uuid of uuids) {
+
+        if ( !ssc.compendiumSourceFromCache(uuid) ) {
+            // Save the tile compendium source in the cache referencing the uuid
+            const source = await fromUuid(uuid);
+            if ( source ) ssc.cacheCompendiumSource(uuid, source);
+        }
+        const source = ssc.compendiumSourceFromCache(uuid);
+        if ( !source ) continue;
+
+        if ( !ssc.spriteFromCache(uuid) ) {
+            // Have Foundry load the texture
+            await TextureLoader.loader.load([source.background.src])
+            ssc.cacheSubSceneSprite(uuid, new PIXI.Sprite(await loadTexture(source.background.src)));
+        }
+
+        if ( !ssc.getSubSceneTile(uuid) ) {
+            // Create a local memory tile for this source.  (not saved to database)
+            const data = {
+                x: 0,
+                y: 0,
+                width: source.width,
+                height: source.height,
+                overhead: false,
+                img: source.background.src,
+                _id: foundry.utils.randomID(16)
+            }
+            const tileDoc = new TileDocument(data, {parent: canvas.scene});
+
+            // Save this tile in the cache referencing both tile.id and the scene uuid, for convenience
+            ssc.setSubSceneCache(tileDoc.id, tileDoc);
+            ssc.setSubSceneCache(uuid, tileDoc);
+        }
+        const tileDoc = ssc.getSubSceneTile(uuid);
+        tileDoc.compendiumSubSceneUUID = uuid;
+
+        const sourcePath = source.getFlag("scene-scroller-maker", ssc.compendiumFlags[3]);
+        tileDoc.imagePath = PolygonMesher.getClipperPathFromPoints(sourcePath);
+
+        // Cache the placeables for this sub-scene
+        cacheInScenePlaceables(source, tileDoc);
+    }
+}
+
+export const debounceGrandChildCache = foundry.utils.debounce( async (uuidArr) => {
+    const missingGrandChildren = new Set();
+    for (const childUuid of uuidArr) {
+        const childSource = await fromUuid(childUuid);
+        const childFlags = childSource.getFlag("scene-scroller-maker", ssc.compendiumFlags[0]);
+        const grandChildrenUUIDs = childFlags.map(s => {
+            return s[ssc.subSceneChildrenFlags[0]];
+        });
+        grandChildrenUUIDs.forEach(u => {
+            if ( !ssc.getSubSceneTile(u) ) missingGrandChildren.add(u);
+        })
+    }
+    cacheSubScene([...missingGrandChildren], {isGrandChild : true});
+}, 1000);
