+++ Lib/Wrap.js
@@ -1,5 +1,4 @@
-import { ModuleName } from "../ss-initialize.js";
-import { log, isScrollerScene } from "./functions.js";
+
 
 /** LibWrappers are initialized on 'init' hook.  (can also work on 'ready' hook) See ss-initialize.js */
